# shared/api/bundles/order.openapi.v1.bundle.yaml
openapi: 3.0.3
info:
  title: OrderService API
  version: 1.0.0
  description: API для работы с заказами
tags:
  - name: Orders
    description: Операции с заказами — получение и обновлен
paths:
  /api/v1/order:
    post:
      summary: Create Order
      operationId: CreateOrder
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_order_request'
      responses:
        '201':
          description: Order successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/create_order_response'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: default case
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/order/{order_uuid}:
    get:
      summary: Get order by UUID
      operationId: GetOrderByUUID
      tags:
        - Orders
      parameters:
        - $ref: '#/components/parameters/order_uuid'
        - $ref: '#/components/parameters/session_uuid'
      responses:
        '200':
          description: Order retrived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order_dto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Other / default error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/order/{order_uuid}/pay:
    post:
      summary: Pay for an order
      operationId: PayOrder
      tags:
        - Orders
      parameters:
        - $ref: '#/components/parameters/order_uuid'
        - $ref: '#/components/parameters/session_uuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pay_order_request'
      responses:
        '200':
          description: Payment successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/pay_order_response'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_server_error'
        default:
          description: Other / default error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
  /api/v1/order/{order_uuid}/cancel:
    post:
      summary: Cancel an order
      operationId: CancelOrder
      tags:
        - Orders
      parameters:
        - $ref: '#/components/parameters/order_uuid'
        - $ref: '#/components/parameters/session_uuid'
      responses:
        '204':
          description: Order successfully cancelled
        '400':
          description: Order bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bad_request_error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/not_found_error'
        '409':
          description: Order already paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/conflict_error'
        default:
          description: Other / default error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/generic_error'
components:
  schemas:
    create_order_request:
      type: object
      required:
        - user_uuid
        - part_uuids
      properties:
        user_uuid:
          type: string
          format: uuid
          description: UUID пользователя
        part_uuids:
          type: array
          items:
            type: string
            format: uuid
          description: Список UUID деталей для заказа
    generic_error:
      type: object
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 404
        message:
          type: string
          description: Описание ошибки
    create_order_response:
      type: object
      required:
        - order_uuid
        - total_price
      properties:
        order_uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор заказа
        total_price:
          type: number
          format: double
          description: Итоговая стоимость
    bad_request_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 400
        message:
          type: string
          description: Описание ошибки
          example: 'Bad request: Invalid parameters format'
    internal_server_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 500
        message:
          type: string
          description: Описание ошибки
          example: Internal server error occurred
    payment_method:
      type: string
      enum:
        - UNKNOWN
        - CARD
        - SBP
        - CREDIT_CARD
      description: Способ оплаты
    order_status:
      type: string
      enum:
        - PENDING_PAYMENT
        - PAID
        - CANCELLED
      description: Статус заказа
    order_dto:
      type: object
      required:
        - order_uuid
        - user_uuid
        - part_uuids
        - total_price
        - status
      properties:
        order_uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор заказа
          example: 0886799f-5898-4236-93e8-ba41f239d825
        user_uuid:
          type: string
          format: uuid
          description: UUID пользователя
          example: 0886799f-5898-4236-93e8-ba41f239d825
        part_uuids:
          type: array
          items:
            type: string
            format: uuid
            description: Список UUID деталей
            example: 0886799f-5898-4236-93e8-ba41f239d825
        total_price:
          type: number
          format: double
          description: Итоговая стоимость
          example: 1554.21
        transaction_uuid:
          type: string
          nullable: true
          description: UUID транзакции (если оплачен)
          example: 0886799f-5898-4236-93e8-ba41f239d825
        payment_method:
          $ref: '#/components/schemas/payment_method'
          nullable: true
        status:
          $ref: '#/components/schemas/order_status'
          nullable: true
    not_found_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 404
        message:
          type: string
          description: Описание ошибки
          example: Order by uuid not found
    pay_order_request:
      type: object
      required:
        - payment_method
      properties:
        payment_method:
          $ref: '#/components/schemas/payment_method'
    pay_order_response:
      type: object
      required:
        - transaction_uuid
      properties:
        transaction_uuid:
          type: string
          format: uuid
          description: Уникальный идентификатор транзакции
    conflict_error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP-код ошибки
          example: 409
        message:
          type: string
          description: Описание ошибки
          example: Conflict error
  parameters:
    order_uuid:
      name: order_uuid
      in: path
      required: true
      description: UUID пользователя, для которого запрашиваются или обновляются данные
      schema:
        type: string
        format: uuid
        example: 0886799f-5898-4236-93e8-ba41f239d825
    session_uuid:
      name: X-Session-Uuid
      in: header
      required: true
      description: UUID сессии пользователя для аутентификации
      schema:
        type: string
        format: uuid
x-ogen:
  target: ./pkg/openapi/order/v1
  package: order_v1
  clean: true
  embed: true